Add files via upload
A number of modifications, including: 264: direction was backwards; reversed it. 285: shortened centerline because end was in the wrong place. 286: lengthened centerline because end was in the wrong place. 402: shortened centerline, because end was in wrong place. 571: moved western end of centerline to the north, to be closer to the node it connects to.