changed node file and new files
nodeUtmNad83.cdl had an extra node 715 that was not part of the grid, so it was moved
gates.cdl contains gate locations. These are not the exact physical locations, but the locations in DSM2.
boundary flows contains points placed at the nodes where boundary flows are applied
boundaryStage contains a point placed at the node where boundary stage is applied.